chore(process-cooling): move types and models to models folder
