+++ src/app/process-cooling-assessment/baseline/baseline-tabs/baseline-tabs.component.ts
@@ -1,5 +1,6 @@
 import { Component, DestroyRef, ElementRef, inject, Signal, ViewChild } from '@angular/core';
-import { ProcessCoolingUiService, SETUP_VIEW_LINKS, ViewLink } from '../../services/process-cooling-ui.service';
+import { ProcessCoolingUiService } from '../../services/process-cooling-ui.service';
+import { SETUP_VIEW_LINKS, ViewLink } from '../../models/views';
 import { ROUTE_TOKENS } from '../../constants/process-cooling-routes';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { ProcessCoolingAssessmentService } from '../../services/process-cooling-assessment.service';
