+++ src/app/process-cooling-assessment/services/process-cooling-ui.service.ts
@@ -9,6 +9,8 @@ import { SystemInformationFormService } from '../system-information/system-infor
 import { ROUTE_TOKENS as WEATHER_ROUTE_TOKENS } from '../../shared/modules/weather-data/models/routes';
 import { SummaryView } from './executive-summary-results.service';
 import { ProfileView } from './system-profile.service';
+import { ProcessCoolingView, SetupView, SystemInformationView, ViewLink } from '../models/views';
+export { ProcessCoolingMainTabString, ProcessCoolingSetupTabString, ProcessCoolingAssessmentTabString, ProcessCoolingView, MainView, SetupView, SystemInformationView, AssessmentView, ReportView, ViewLink, MAIN_VIEW_LINKS, SETUP_VIEW_LINKS, SYSTEM_INFORMATION_VIEW_LINKS, ASSESSMENT_VIEW_LINKS, REPORT_VIEW_LINKS } from '../models/views';
 
 @Injectable()
 export class ProcessCoolingUiService {
@@ -88,6 +90,7 @@ export class ProcessCoolingUiService {
   // todo 8129 optimization - called on every rerender. 
   // todo this pattern should be changed or improved when we get concrete information on what is required for results calculation. memoize, or map/array of vals to Observable
   canVisitSteppedView(steppedRouteIndex: number): boolean {
+    console.log('canVisitSteppedView');
     const processCooling = this.processCoolingAssessmentService.processCoolingSignal();
     const settings = this.processCoolingAssessmentService.settingsSignal();
 
@@ -292,167 +295,3 @@ export class ProcessCoolingUiService {
   }
 
 }
-
-// todo move below out to models/types folder
-export type ProcessCoolingMainTabString = 'baseline' | 'assessment' | 'diagram' | 'report' | 'calculators';
-export type ProcessCoolingSetupTabString = 'assessment-settings' | 'system-information' | 'inventory' | 'operating-schedule' | 'load-schedule';
-export type ProcessCoolingAssessmentTabString = 'explore-opportunities';
-export type ProcessCoolingView = SetupView | MainView | ReportView | AssessmentView | SystemInformationView;
-
-export enum MainView {
-  BASELINE = 'baseline',
-  ASSESSMENT = 'assessment',
-  REPORT = 'report',
-}
-
-export enum SetupView {
-  ASSESSMENT_SETTINGS = 'assessment-settings',
-  SYSTEM_INFORMATION = 'system-information',
-  INVENTORY = 'chiller-inventory',
-  OPERATING_SCHEDULE = 'operating-schedule',
-  LOAD_SCHEDULE = 'load-schedule',
-}
-
-export enum SystemInformationView {
-  OPERATIONS = 'operations',
-  WEATHER = 'weather',
-  PUMP = 'pump',
-  CONDENSER_COOLING_SYSTEM_INPUT = 'condenser-cooling-system',
-  TOWER = 'tower',
-}
-
-export enum AssessmentView {
-  EXPLORE_OPPORTUNITIES = 'explore-opportunities',
-}
-
-export enum ReportView {
-  EXECUTIVE_SUMMARY = 'executive-summary',
-  PERFORMANCE_PROFILE = 'performance-profile',
-  PUMP_SUMMARY = 'pump-summary',
-  TOWER_SUMMARY = 'tower-summary',
-  INPUT_SUMMARY = 'input-summary',
-  GRAPHS = 'graphs',
-  SYSTEM_PROFILE = "system-profile",
-  FACILITY_INFO = 'facility-info',
-}
-
-
-export interface ViewLink {
-  view: ProcessCoolingView;
-  label: string;
-  param?: string | number;
-  meta?: {
-    disabled?: boolean;
-  };
-}
-
-export const MAIN_VIEW_LINKS: ViewLink[] = [
-  {
-    view: MainView.BASELINE,
-    label: 'Baseline',
-  },
-  {
-    view: MainView.ASSESSMENT,
-    label: 'Assessment',
-  },
-  {
-    view: MainView.REPORT,
-    label: 'Report',
-  },
-];
-
-export const SETUP_VIEW_LINKS: ViewLink[] = [
-  {
-    view: SetupView.ASSESSMENT_SETTINGS,
-    label: 'Assessment Settings',
-  },
-  {
-    view: SetupView.SYSTEM_INFORMATION,
-    label: 'System Information',
-  },
-  {
-    view: SetupView.INVENTORY,
-    label: 'Chiller Inventory',
-  },
-  {
-    view: SetupView.OPERATING_SCHEDULE,
-    label: 'Operating Schedule',
-  },
-  {
-    view: SetupView.LOAD_SCHEDULE,
-    label: 'Load Schedule',
-  },
-]
-
-// todo get is route available via filter pipe
-export const SYSTEM_INFORMATION_VIEW_LINKS: ViewLink[] = [
-  {
-    view: SystemInformationView.OPERATIONS,
-    label: 'Operations',
-  },
-  {
-    view: SystemInformationView.WEATHER,
-    label: 'Weather',
-  },
-  {
-    view: SystemInformationView.PUMP,
-    label: 'Pump',
-  },
-  {
-    view: SystemInformationView.CONDENSER_COOLING_SYSTEM_INPUT,
-    label: 'Condenser Cooling'
-  },
-  {
-    view: SystemInformationView.TOWER,
-    label: 'Tower',
-  },
-
-]
-
-export const ASSESSMENT_VIEW_LINKS: ViewLink[] = [
-  {
-    view: AssessmentView.EXPLORE_OPPORTUNITIES,
-    label: 'Explore Opportunities',
-  },
-]
-
-export const REPORT_VIEW_LINKS: ViewLink[] = [
-  {
-    view: ReportView.EXECUTIVE_SUMMARY,
-    label: 'Executive Summary',
-  },
-  {
-    view: ReportView.PERFORMANCE_PROFILE,
-    label: 'Performance Profile',
-    // meta: {
-    //   disabled: true
-    // }
-  },
-  {
-    view: ReportView.SYSTEM_PROFILE,
-    label: 'System Profile',
-  },
-  {
-    view: ReportView.PUMP_SUMMARY,
-    label: 'Pump Summary',
-  },
-  {
-    view: ReportView.TOWER_SUMMARY,
-    label: 'Tower Summary',
-  },
-  {
-    view: ReportView.INPUT_SUMMARY,
-    label: 'Input Summary',
-  },
-  {
-    view: ReportView.FACILITY_INFO,
-    label: 'Facility Info',
-  },
-  {
-    view: ReportView.GRAPHS,
-    label: 'Graphs',
-    meta: {
-      disabled: true
-    }
-  },
-]
