+++ src/app/process-cooling-assessment/system-information/system-information.component.ts
@@ -1,5 +1,6 @@
 import { Component, DestroyRef, inject, Signal } from '@angular/core';
-import { ProcessCoolingUiService, SYSTEM_INFORMATION_VIEW_LINKS, ViewLink } from '../services/process-cooling-ui.service';
+import { ProcessCoolingUiService } from '../services/process-cooling-ui.service';
+import { SYSTEM_INFORMATION_VIEW_LINKS, ViewLink } from '../models/views';
 import { ROUTE_TOKENS } from '../constants/process-cooling-routes';
 import { ProcessCoolingAssessmentService } from '../services/process-cooling-assessment.service';
 import { CondenserCoolingMethod } from '../../shared/models/process-cooling-assessment';
